V1.0

+++ src/app.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Product } from 'models/product.entity';
@@ -15,22 +15,25 @@ export class AppService {
     return this.productRepository.find();
   }
 
-  findOne(id): Promise<Product> {
-    return this.productRepository.findOne(id);
-  }
-
-  create(product: Product): Promise<Product> {
+async findOne(id: number): Promise<Product | undefined> {
+  return this.productRepository.findOne({ where: { id } });
+}
+  
+create(product: Product): Promise<Product> {
     return this.productRepository.save(product);
   }
 
-  async update(id, product: Product): Promise<Product> {
+  async update(id: number, product: Product): Promise<Product> {
     await this.productRepository.update(id, product);
-    return this.productRepository.findOne(id);
+    return this.productRepository.findOne({ where: { id } });
   }
 
-  async delete(id: number): Promise<void> {
-    await this.productRepository.delete(id);
+async delete(id: number): Promise<void> {
+  const product = await this.productRepository.findOne({ where: { id } });
+  if (!product) {
+    throw new NotFoundException(`Producto con ID ${id} no encontrado.`);
   }
-
+  await this.productRepository.delete(id);
+}
 
 }
